+++ Runtime/RHI/RHI_Pipeline.cpp
@@ -355,7 +355,7 @@ namespace Directus
 		if (m_viewportDirty)
 		{
 			m_rhiDevice->Set_Viewport(m_viewport);
-			Settings::Get().Viewport_Set((int)m_viewport->GetWidth(), (int)m_viewport->GetHeight());
+			Settings::Get().Viewport_Set(m_viewport->GetWidth(), m_viewport->GetHeight());
 			m_viewportDirty = false;
 		}
